+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-Unreleased
-----------
+0.11.0
+------
 - Added support for bracket-style orders
 - Added `From` implementation for `asset::Symbol` type
 - Added support for almost-default construction of various request types